+++ epi/src/lib.rs
@@ -304,6 +304,18 @@ pub trait TextureAllocator {
     fn free(&mut self, id: egui::TextureId);
 }
 
+/// Abstraction for platform dependent texture reference
+pub trait NativeTexture {
+    type Texture;
+
+    /// Bind native texture to egui texture
+    fn register_native_texture(&mut self, native: Self::Texture) -> egui::TextureId;
+
+    /// Change id's actual pointing texture
+    /// only for user texture
+    fn replace_native_texture(&mut self, id: egui::TextureId, replacing: Self::Texture);
+}
+
 /// How to signal the [`egui`] integration that a repaint is required.
 pub trait RepaintSignal: Send + Sync {
     /// This signals the [`egui`] integration that a repaint is required.