+++ egui_web/CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to the `egui_web` integration will be noted in this file.
 
 
 ## Unreleased
-* Remove "http" feature (use https://github.com/emilk/ehttp instead!).
 
+### Added
+* Remove "http" feature (use https://github.com/emilk/ehttp instead!).
+* `epi::NativeTexture` trait for webgl1 webgl2 painter
+* Deprecate `Painter::register_webgl_texture` 
 
 ## 0.14.1 - 2021-08-28